+++ examples/signal.py
@@ -1,0 +1,10 @@
+from scipy import signal
+
+
+print(signal.gaussian(51, std=7))
+
+
+print(signal.gaussian(5, std=3))
+
+
+print(signal.gaussian(3, std=1))