+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 Major/breaking changes listed.
 
+## 0.26
+### Added
+* rolling (rolling windows of computed functions - math operation)
+* dataset/dssort-by
+* dataset/ds-take-nth
+
 ## 0.22
 ### Added
 * PCA